pyproject.toml
[project]
name = "parsing_benchmark"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
  "rust_pest",
  "python_lark",
]

[tool.uv.sources]
rust_pest = { workspace = true }
python_lark = { workspace = true }

[tool.uv.workspace]
members = ["rust_pest", "python_lark"]

[dependency-groups]
dev = [
    "django-stubs>=5.1.3",
    "mypy>=1.15.0",
    "pytest>=8.3.5",
    "pytest-benchmark[histogram]>=5.1.0",
    "pytest-icdiff>=0.9",
    "ruff>=0.9.9",
]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["ALL"]
ignore = ["D", "S101"]
